Refactor stage progression into GameEngine
